chore(rig-memory): move inline test modules to sibling files
